+++ InventoryTools/Logic/ItemRenderers/ItemSubmarineDropSourceRenderer.cs
@@ -28,6 +28,12 @@ public class ItemSubmarineDropSourceRenderer : ItemInfoRenderer<ItemSubmarineDro
     public override string HelpText => "Can the item be earned from a submarine exploration route?";
     public override bool ShouldGroup => true;
 
+    public override Func<ItemSource, (Type, uint)>? RelatedType => source =>
+    {
+        var asSource = AsSource(source);
+        return (asSource.SubmarineExploration.GetType(), asSource.SubmarineExploration.RowId);
+    };
+
     public override Func<ItemSource, List<MessageBase>?>? OnClick => source =>
     {
         var asSource = AsSource(source);